requirements.txt
pytest==5.3.5
requests==2.22.0
jsonschema==3.2.0
allure-pytest==2.8.18
allure-python-commons==2.8.18
